@@ -1,2 +1,2 @@
-@javax.xml.bind.annotation.XmlSchema( elementFormDefault = javax.xml.bind.annotation.XmlNsForm.QUALIFIED, namespace="http://www.someuri" )
+@jakarta.xml.bind.annotation.XmlSchema( elementFormDefault = jakarta.xml.bind.annotation.XmlNsForm.QUALIFIED, namespace="http://www.someuri" )
 package org.acme;
