@@ -4,5 +4,5 @@
 @XmlSchema(namespace = "example", elementFormDefault = XmlNsForm.QUALIFIED)
 package org.testing.xml;
 
-import javax.xml.bind.annotation.XmlSchema;
-import javax.xml.bind.annotation.XmlNsForm;
+import jakarta.xml.bind.annotation.XmlSchema;
+import jakarta.xml.bind.annotation.XmlNsForm;
