@@ -43,7 +43,7 @@ File buildLog = new File(basedir, 'build.log')
 def logPattern = "\\| \\[\\p{Digit}+\\]: ";
 def sep = Pattern.quote(System.getProperty("file.separator"));
 Pattern expectedTargetPattern  = Pattern.compile(logPattern + "\\-target");
-Pattern expectedTargetVersionPattern = Pattern.compile(logPattern + "2.0");
+Pattern expectedTargetVersionPattern = Pattern.compile(logPattern + "3.0");
 Pattern expectedExtensionPattern = Pattern.compile(logPattern + "\\-extension");
 Pattern expectedEpisodePattern = Pattern.compile(logPattern + "\\-episode");
 Pattern expectedEpisodeFilePattern = Pattern.compile(logPattern +
